requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
jinja2==3.1.2 